vector_databases/requirements.txt
# Embeddings
sentence-transformers>=2.2.2

# Vector DBs and libs
chromadb>=0.4.0
faiss-cpu>=1.7.4
lancedb>=0.7.0

# Utilities
numpy>=1.24
pandas>=2.1.0
tqdm>=4.65
